streamlit-v1/requirements.txt
streamlit==1.40.2
streamlit-autorefresh==1.0.1
extra-streamlit-components==0.1.71
pandas==2.2.0
numpy==1.26.4
plotly==5.24.1
scipy==1.14.0
requests==2.32.3
yfinance==0.2.66
pytz==2024.2
bcrypt==4.2.0
beautifulsoup4==4.13.2
lxml==5.3.0
python-dotenv==1.0.1
gunicorn==23.0.0
